requirements.txt
anyio=4.2.0
appnope=0.1.2
argon2-cffi=21.3.0
argon2-cffi-bindings=21.2.0
asttokens=2.0.5
async-lru=2.0.4
attrs=23.1.0
Babel=2.11.0
beautifulsoup4=4.12.3
bleach=4.1.0
Bottleneck=1.3.7
branca=0.6.0
certifi=2024.8.30
cffi=1.16.0
charset-normalizer=2.0.4
click=8.1.7
click-plugins=1.1.1
cligj=0.7.2
comm=0.2.1
contourpy=1.2.0
cycler=0.11.0
debugpy=1.6.7
decorator=5.1.1
defusedxml=0.7.1
executing=0.8.3
fastjsonschema=2.16.2
fiona=1.9.5
folium=0.14.0
fonttools=4.51.0
GDAL=3.6.2
geopandas=0.14.2
idna=3.4
imbalanced-learn=0.11.0
ipykernel=6.28.0
ipython=8.20.0
jedi=0.18.1
Jinja2=3.1.3
joblib=1.2.0
json5=0.9.6
jsonschema=4.19.2
jsonschema-specifications=2023.7.1
jupyter_client=8.6.0
jupyter_core=5.5.0
jupyter-events=0.8.0
jupyter-lsp=2.2.0
jupyter_server=2.10.0
jupyter_server_terminals=0.4.4
jupyterlab=4.0.11
jupyterlab-pygments=0.1.2
jupyterlab_server=2.25.1
keras=3.4.1
Keras-Preprocessing=1.1.2
kiwisolver=1.4.4
lxml=5.2.1
mapclassify=2.5.0
MarkupSafe=2.1.3
matplotlib=3.8.0
matplotlib-inline=0.1.6
mistune=2.0.4
nbclient=0.8.0
nbconvert=7.10.0
nbformat=5.9.2
nest-asyncio=1.6.0
networkx=3.1
notebook=7.0.8
notebook_shim=0.2.3
numexpr=2.8.7
numpy=1.26.4
overrides=7.4.0
packaging=23.2
pandas=2.2.1
pandocfilters=1.5.0
parso=0.8.3
patsy=0.5.3
pexpect=4.8.0
pillow=10.2.0
pip=23.3.1
platformdirs=3.10.0
plotly=5.19.0
pooch=1.7.0
prometheus-client=0.14.1
prompt-toolkit=3.0.43
psutil=5.9.0
ptyprocess=0.7.0
pure-eval=0.2.2
pycparser=2.21
Pygments=2.15.1
pyparsing=3.0.9
pyproj=3.6.1
python-dateutil=2.8.2
python-json-logger=2.0.7
pytz=2023.3.post1
PyYAML=6.0.1
pyzmq=25.1.2
referencing=0.30.2
requests=2.31.0
rfc3339-validator=0.1.4
rfc3986-validator=0.1.1
rpds-py=0.10.6
Rtree=1.0.1
scikit-learn=1.1.3
scikit-plot=0.3.7
scipy=1.10.0
seaborn=0.12.2
Send2Trash=1.8.2
setuptools=68.2.2
shapely=2.0.1
six=1.16.0
sklearn-pandas=2.2.0
sniffio=1.3.0
soupsieve=2.5
stack-data=0.2.0
statsmodels=0.14.0
streamlit=1.39.0
streamlit-antd-components=0.3.2
tenacity=8.2.2
terminado=0.17.1
threadpoolctl=2.2.0
tinycss2=1.2.1
tornado=6.3.3
traitlets=5.7.1
typing_extensions=4.9.0
tzdata=2023.3
unicodedata2=15.1.0
urllib3=2.1.0
wcwidth=0.2.5
webencodings=0.5.1
websocket-client=0.58.0
wheel=0.41.2
xgboost=1.7.3
xyzservices=2022.9.0
pydeck==3.1.3
